zipfer 0.1.0

+++ src/root.zig
@@ -1,4 +1,4 @@
-pub const Zipfer = ZipferImpl(f32);
+pub const Zipfer = ZipferImpl(f64);
 pub const ZipferImpl = @import("zipfer.zig").ZipferImpl;
 
 const std = @import("std");
